Moved DatePicker and PoweredByCalendso to seperate components
